# stage1_gateway_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ForgeAI - Api Gateway",
    "description": "API Gateway single public entry point providing authentication, resilience, and metrics.\n\n*Part of the ForgeAI Local AI Software Engineering Platform.*",
    "version": "0.1.0",
    "x-service-name": "api-gateway"
  },
  "paths": {
...
    "/api/v1/git/status": {
      "get": {
        "tags": [
          "Git Version Control"
        ],
        "summary": "Get Git Repository Status",
        "description": "Returns current branch, staged, modified, and untracked files.",
        "operationId": "get_status_api_v1_git_status_get",
        "parameters": [
          {
            "name": "repository_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Repository Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Git status retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_GitStatusResponse_"
                }
              }
            }
          },
          "503": {
            "description": "Git Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/git/diff": {
      "get": {
        "tags": [
          "Git Version Control"
        ],
        "summary": "Get Git Unified Diff",
        "description": "Returns unified git diff text for working tree modifications.",
        "operationId": "get_diff_api_v1_git_diff_get",
        "parameters": [
          {
            "name": "repository_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Repository Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unified diff retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_GitDiffResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/git/log": {
      "get": {
        "tags": [
          "Git Version Control"
        ],
        "summary": "Get Git Commit History Log",
        "description": "Returns recent commit history objects.",
        "operationId": "get_log_api_v1_git_log_get",
        "parameters": [
          {
            "name": "repository_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Repository Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Git log retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_GitLogResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/git/branches": {
      "post": {
        "tags": [
          "Git Version Control"
        ],
        "summary": "Create Git Branch",
        "description": "Creates a new local git branch in workspace.",
        "operationId": "create_branch_api_v1_git_branches_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitCreateBranchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Branch created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_GitActionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/git/checkout": {
      "post": {
        "tags": [
          "Git Version Control"
        ],
        "summary": "Checkout Git Branch or Commit",
        "description": "Checks out specified target branch or commit SHA.",
        "operationId": "checkout_target_api_v1_git_checkout_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitCheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Target checked out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_GitActionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/git/commit": {
      "post": {
        "tags": [
          "Git Version Control"
        ],
        "summary": "Create Git Commit",
        "description": "Stages and commits working tree modifications.",
        "operationId": "create_commit_api_v1_git_commit_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitCommitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Commit created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_GitActionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/git/restore": {
      "post": {
        "tags": [
          "Git Version Control"
        ],
        "summary": "Discard Working Directory Modifications",
        "description": "Reverts modified/deleted files back to HEAD commit state.",
        "operationId": "restore_files_api_v1_git_restore_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitRestoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Files restored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_GitActionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/git/stage": {
      "post": {
        "tags": [
          "Git Version Control"
        ],
        "summary": "Stage Working Tree Files",
        "description": "Stages specified files into the Git index (git add).",
        "operationId": "stage_files_api_v1_git_stage_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitStageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Files staged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_GitActionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/git/unstage": {
      "post": {
        "tags": [
          "Git Version Control"
        ],
        "summary": "Unstage Staged Files",
        "description": "Unstages specified files from the Git index (git restore --staged).",
        "operationId": "unstage_files_api_v1_git_unstage_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitUnstageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Files unstaged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_GitActionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetails": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable error code"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Additional context or validation errors"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "ErrorDetails",
        "description": "Detailed error info structure."
      },
      "ErrorResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "False for failed requests",
            "default": false
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Correlation ID for request tracing",
            "default": "-"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "Distributed trace ID",
            "default": "-"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 UTC timestamp"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable error code",
            "default": "INTERNAL_ERROR"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error description"
          },
          "service": {
            "type": "string",
            "title": "Service",
            "description": "Originating service name",
            "default": "api-gateway"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetails",
            "description": "Structured error details"
          }
        },
        "type": "object",
        "required": [
          "message",
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Unified error response wrapper conforming to production contract."
      },
...
      "GitCheckoutRequest": {
        "properties": {
          "target": {
            "type": "string",
            "title": "Target",
            "description": "Branch name or commit hash to checkout",
            "example": "main"
          },
          "repository_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Id",
            "description": "Optional target repository ID"
          }
        },
        "type": "object",
        "required": [
          "target"
        ],
        "title": "GitCheckoutRequest",
        "description": "Request payload to checkout an existing branch or commit SHA."
      },
...
      "GitCreateBranchRequest": {
        "properties": {
          "branch_name": {
            "type": "string",
            "title": "Branch Name",
            "description": "New branch name",
            "example": "feature/auth-rate-limit"
          },
          "checkout": {
            "type": "boolean",
            "title": "Checkout",
            "description": "Whether to checkout the branch immediately",
            "default": true
          },
          "repository_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Id",
            "description": "Optional target repository ID"
          }
        },
        "type": "object",
        "required": [
          "branch_name"
        ],
        "title": "GitCreateBranchRequest",
        "description": "Request payload to create a new git branch."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SuccessResponse_GitActionResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "True for successful requests",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Correlation ID for request tracing",
            "default": "-"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "Distributed trace ID",
            "default": "-"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 UTC timestamp"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GitActionResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payload data"
          }
        },
        "type": "object",
        "title": "SuccessResponse[GitActionResponse]"
      },
      "SuccessResponse_GitDiffResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "True for successful requests",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Correlation ID for request tracing",
            "default": "-"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "Distributed trace ID",
            "default": "-"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 UTC timestamp"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GitDiffResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payload data"
          }
        },
        "type": "object",
        "title": "SuccessResponse[GitDiffResponse]"
      },
      "SuccessResponse_GitLogResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "True for successful requests",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Correlation ID for request tracing",
            "default": "-"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "Distributed trace ID",
            "default": "-"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 UTC timestamp"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GitLogResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payload data"
          }
        },
        "type": "object",
        "title": "SuccessResponse[GitLogResponse]"
      },
      "SuccessResponse_GitStatusResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "True for successful requests",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Correlation ID for request tracing",
            "default": "-"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "Distributed trace ID",
            "default": "-"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 UTC timestamp"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GitStatusResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payload data"
          }
        },
        "type": "object",
        "title": "SuccessResponse[GitStatusResponse]"
      },
...
    }
  }
}